==> 01_data_preparation.Rmd <==
---
title: "ICPS/Lariboisiere - ML CMR CCTA"
author: "Solenn Toupin, PhD"
date: "08/03/2024"
output:
  html_document:
    toc: yes
    toc_depth: 3
    number_sections: no
    theme: united
    highlight: tango
  pdf_document: default
  html_notebook:
    toc: yes
    toc_float: yes
  editor_options:
    markdown:
      wrap: sentence
  word_document:
    toc: yes
    toc_depth: '3'
---

```{r, echo = FALSE, warning= FALSE}
library(tidyverse)
library(dplyr)
library(gtsummary)
library(officer)
library(flextable)
knitr::asis_output("\U00B1")
library(survival)
library(survminer)
library(ggplot2)
library(readr)
library("readxl")
library(partykit)
library(mfp)
library(survIDINRI)
library(glmnet)
library(survival)
library(caret)
library(Hmisc)

```


```{r, include = FALSE, echo = FALSE, warning = FALSE}
data <- read_excel("/Users/solenntoupin/Documents/Projets_R/ML_CMR_CCTA_mise_au_propre/data/data.xlsx")

# Rename the variables
names(data)[1] <- "ID"
names(data)[2] <- "Centre"
names(data)[3] <- "Ischemia"
names(data)[4] <- "Revascularization"
names(data)[5] <- "Gender"
names(data)[6] <- "Followup_Time"
names(data)[7] <- "Competing_Event"
names(data)[8] <- "Time_Competing_Event"
names(data)[9] <- "MACE"
names(data)[10] <- "Time_MACE"
names(data)[11] <- "Death"
names(data)[12] <- "Time_Death"
names(data)[13] <- "CV_Death"
names(data)[14] <- "Time_CV_Death"
names(data)[15] <- "MI"
names(data)[16] <- "Time_MI"
names(data)[17] <- "Age"
names(data)[18] <- "Cardiac_Rhythm"
names(data)[19] <- "BMI"
names(data)[20] <- "Obesity"
names(data)[21] <- "BSA"
names(data)[22] <- "Dyslipidemia"
names(data)[23] <- "Diabetes"
names(data)[24] <- "Hypertension"
names(data)[25] <- "Smoking"
names(data)[26] <- "Family_History_CAD"
names(data)[27] <- "Renal_Failure"
names(data)[28] <- "Stroke"
names(data)[29] <- "Pacemaker"
names(data)[30] <- "Peripheral_Atheroma"
names(data)[31] <- "History_Hospit_HF"
names(data)[32] <- "Symptomatic_Angina"
names(data)[33] <- "Dyspnea"
names(data)[34] <- "LVEF"
names(data)[35] <- "LVEDVi"
names(data)[36] <- "LVESVi"
names(data)[37] <- "RV_Dysfunction"
names(data)[38] <- "LGE"
names(data)[39] <- "Viability_if_LGE"
names(data)[40] <- "Ischemia_Segments"
names(data)[41] <- "LGE_Segments"
names(data)[42] <- "Score_C_CMR_10"
names(data)[43] <- "Score_ESC"
names(data)[44] <- "Score_QRISK3"
names(data)[45] <- "score_FRS"
names(data)[46] <- "Segments_Stenosis_Sup_70percent"
names(data)[47] <- "Segments_Stenosis_Sup_50percent"
names(data)[48] <- "Proximal_Segments_Stenosis_Sup_50percent"
names(data)[49] <- "Segments_Noncalcified_Plaques"
names(data)[50] <- "Segments_Mixed_Plaques"
names(data)[51] <- "Segments_Calcified_Plaques"
names(data)[52] <- "Proximal_Segments_Stenosis_Sup_70percent"
names(data)[53] <- "Vessels_Obstructive_CAD"

data$Segments_Stenosis_Sup_70percent <- as.numeric(data$Segments_Stenosis_Sup_70percent)
data$Segments_Stenosis_Sup_50percent <- as.numeric(data$Segments_Stenosis_Sup_50percent)
data$Proximal_Segments_Stenosis_Sup_50percent <- as.numeric(data$Proximal_Segments_Stenosis_Sup_50percent)
data$Segments_Noncalcified_Plaques <- as.numeric(data$Segments_Noncalcified_Plaques)
data$Segments_Mixed_Plaques <- as.numeric(data$Segments_Mixed_Plaques)
data$Segments_Calcified_Plaques <- as.numeric(data$Segments_Calcified_Plaques)
data$Proximal_Segments_Stenosis_Sup_70percent <- as.numeric(data$Proximal_Segments_Stenosis_Sup_70percent)
data$Vessels_Obstructive_CAD <- as.numeric(data$Vessels_Obstructive_CAD)

data <- data %>% 
        mutate(Time_MACE_years = Time_MACE/12,
               Segments_Calcified_or_Mixed_Plaques = Segments_Calcified_Plaques + Segments_Mixed_Plaques,
               Score_SIS = Segments_Stenosis_Sup_70percent + Segments_Stenosis_Sup_50percent,
               Score_SIS_2 = Segments_Noncalcified_Plaques + Segments_Mixed_Plaques + Segments_Calcified_Plaques,
               Score_SSS = 2*Segments_Stenosis_Sup_50percent + 3*Segments_Stenosis_Sup_70percent)

Score_ML_CMR_age <- (data$Age/100)^2*2.23
Score_ML_CMR_gender <- ifelse(data$Gender == 1, 0, -0.65)
Score_ML_CMR_BMI <- (data$BMI/100)^2*1.27
Score_ML_CMR_renal <- ifelse(data$Renal_Failure == 1, 1.02, 0)
Score_ML_CMR_peripheral <- ifelse(data$Peripheral_Atheroma == 1,0.39,0)
Score_ML_CMR_diabetes <- ifelse(data$Diabetes == 1, 0.52, 0)
Score_ML_CMR_LVEF <- (data$LVEF/100)^3*3.49
Score_ML_CMR_LVEDV <- (data$LVEDVi/100)*0.52
Score_ML_CMR_Ischemia <- ifelse(data$Ischemia_Segments == 1 | data$Ischemia_Segments == 2,-0.24,
                                ifelse(data$Ischemia_Segments >= 6, 2.35,
                                       ifelse(data$Ischemia_Segments == 0,0, 0.55)))

Score_ML_CMR_LGE <- ifelse(data$LGE_Segments == 1 | data$LGE_Segments == 2,-0.32,
                                ifelse(data$LGE_Segments >= 6, 1.37,
                                       ifelse(data$LGE_Segments == 0,0, 0.68)))

Score_ML_CMR <- Score_ML_CMR_age + Score_ML_CMR_gender + Score_ML_CMR_BMI + Score_ML_CMR_renal + Score_ML_CMR_peripheral + Score_ML_CMR_diabetes -
        Score_ML_CMR_LVEF - Score_ML_CMR_LVEDV + Score_ML_CMR_Ischemia + Score_ML_CMR_LGE

library(scales)
Score_ML_CMR_0_10 <- rescale(Score_ML_CMR, to = c(0, 10), from = c(-4.17009175,6.881102656))

data <- data %>% 
        mutate(Score_ML_CMR = Score_ML_CMR_0_10)

```

# Descriptive analysis {.tabset}

## Table 1 stratified by center (ICPS and Lariboisiere)

```{r, echo = FALSE, warning = FALSE}

list("style_number-arg:big.mark" = "") %>%
  set_gtsummary_theme()

varlist=c("Centre","Age","Gender","BMI","Obesity","BSA","Diabetes","Hypertension","Dyslipidemia","Smoking","Family_History_CAD","Renal_Failure","History_Hospit_HF","Pacemaker","Stroke","Peripheral_Atheroma","Symptomatic_Angina","Dyspnea","Segments_Stenosis_Sup_70percent","Segments_Stenosis_Sup_50percent","Proximal_Segments_Stenosis_Sup_50percent","Proximal_Segments_Stenosis_Sup_70percent","Segments_Calcified_Plaques","Segments_Noncalcified_Plaques","Segments_Mixed_Plaques","Segments_Calcified_or_Mixed_Plaques","Vessels_Obstructive_CAD","LVEF","LVEDVi","LVESVi","RV_Dysfunction","Ischemia","Ischemia_Segments","LGE","LGE_Segments","Death","MACE","CV_Death","MI","Revascularization")

data %>% 
  dplyr::select(all_of(varlist)) %>%  # keep only the columns of interest
  tbl_summary(by = Centre,
    label = list(),
    type = list(c("Segments_Stenosis_Sup_70percent","Segments_Stenosis_Sup_50percent","Proximal_Segments_Stenosis_Sup_50percent","Proximal_Segments_Stenosis_Sup_70percent","Segments_Calcified_Plaques","Segments_Noncalcified_Plaques","Segments_Mixed_Plaques","Segments_Calcified_or_Mixed_Plaques","Ischemia_Segments","LGE_Segments") ~ "continuous"),
    value = list(),
    digits = list(all_categorical() ~ c(0, 1)),
    statistic = list(all_continuous() ~ paste0("{mean}"," ± ","{sd}"))) %>%
  add_p(pvalue_fun = ~style_pvalue(.x, digits = 3)) %>%
  bold_p() %>%
  add_overall() %>%
  modify_header(stat_0 ~ "**Overall\nN = {N}**", stat_1 ~ "Centre 1 - ICPS \nN = {n}", stat_2 ~ "Centre 2 - Lariboisiere \nN = {n}") %>%
  as_flex_table()

#save_as_docx(Table1, path = "C:/UserData/Z003C6UX/Documents/-- R-PROGRAMMATION/ML_CMR_CCTA_final/output/Table1-stratified_by_centre.docx")

```

## Table 1 stratified by MACE

```{r, echo = FALSE, warning = FALSE}

list("style_number-arg:big.mark" = "") %>%
  set_gtsummary_theme()

varlist=c("Age","Gender","BMI","Obesity","BSA","Diabetes","Hypertension","Dyslipidemia","Smoking","Family_History_CAD","Renal_Failure","History_Hospit_HF","Pacemaker","Stroke","Peripheral_Atheroma","Symptomatic_Angina","Dyspnea","Segments_Stenosis_Sup_70percent","Segments_Stenosis_Sup_50percent","Proximal_Segments_Stenosis_Sup_50percent","Proximal_Segments_Stenosis_Sup_70percent","Segments_Calcified_Plaques","Segments_Noncalcified_Plaques","Segments_Mixed_Plaques","Segments_Calcified_or_Mixed_Plaques","Vessels_Obstructive_CAD","LVEF","LVEDVi","LVESVi","RV_Dysfunction","Ischemia","Ischemia_Segments","LGE","LGE_Segments","MACE","Revascularization")

data %>% 
  select(all_of(varlist)) %>%  # keep only the columns of interest
  tbl_summary(by = MACE,
    label = list(),
    type = list(c("Segments_Stenosis_Sup_70percent","Segments_Stenosis_Sup_50percent","Proximal_Segments_Stenosis_Sup_50percent","Proximal_Segments_Stenosis_Sup_70percent","Segments_Calcified_Plaques","Segments_Noncalcified_Plaques","Segments_Mixed_Plaques","Segments_Calcified_or_Mixed_Plaques","Ischemia_Segments","LGE_Segments") ~ "continuous"),
    value = list(),
    digits = list(all_categorical() ~ c(0, 1)),
    statistic = list(all_continuous() ~ paste0("{mean}"," ± ","{sd}"))) %>%
  add_p(pvalue_fun = ~style_pvalue(.x, digits = 3)) %>%
  bold_p() %>%
  add_overall() %>%
  modify_header(stat_0 ~ "**Overall\nN = {N}**", stat_1 ~ "Without MACE \nN = {n}", stat_2 ~ "With MACE \nN = {n}") %>%
  as_flex_table()

#save_as_docx(Table1, path = "C:/UserData/Z003C6UX/Documents/-- R-PROGRAMMATION/ML_CMR_CCTA_final/output/Table1-stratified_by_MACE.docx")

```


# Split into Train, Test, Validation (Lariboisiere) datasets 

```{r}

Lariboisiere <- filter(data,Centre==2)
data_ICPS <- filter(data,Centre==1)

data_ICPS_select <- data_ICPS %>%
  dplyr::select(-c("ID","Centre","Competing_Event","Time_Competing_Event","MI","Time_MI","Death","Time_Death","CV_Death","Time_CV_Death"))
data_Validation <- Lariboisiere %>%
  dplyr::select(-c("ID","Centre","Competing_Event","Time_Competing_Event","MI","Time_MI","Death","Time_Death","CV_Death","Time_CV_Death"))

set.seed(45)
trainIndex <- createDataPartition(factor(data_ICPS_select$MACE), p = .7, 
                                  list = FALSE)

data_Train <- data_ICPS_select[trainIndex,]
data_Test  <- data_ICPS_select[-trainIndex,]
write.csv(data_Train,'/Users/solenntoupin/Documents/Projets_R/ML_CMR_CCTA_mise_au_propre/data/Dataset_Train.csv')
write.csv(data_Test,'/Users/solenntoupin/Documents/Projets_R/ML_CMR_CCTA_mise_au_propre/data/Dataset_Test.csv')
write.csv(data_Validation,'/Users/solenntoupin/Documents/Projets_R/ML_CMR_CCTA_mise_au_propre/data/Dataset_Validation.csv')

```

# Addition of a 2nd external Validation : American Hospital of Paris
```{r}
data_Validation_2 <- read_excel("/Users/solenntoupin/Documents/Projets_R/ML_CMR_CCTA_mise_au_propre/data/data_validation_2.xlsx")

names(data_Validation_2)[1] <- "ID"
names(data_Validation_2)[2] <- "MACE"
names(data_Validation_2)[3] <- "Time_MACE"
names(data_Validation_2)[4] <- "LVEF"
names(data_Validation_2)[5] <- "Ischemia_Segments"
names(data_Validation_2)[6] <- "LGE_Segments"
names(data_Validation_2)[7] <- "Proximal_Segments_Stenosis_Sup_50percent"
names(data_Validation_2)[8] <- "Segments_Noncalcified_Plaques"
names(data_Validation_2)[9] <- "Vessels_Obstructive_CAD"

write.csv(data_Validation_2,'/Users/solenntoupin/Documents/Projets_R/ML_CMR_CCTA_mise_au_propre/data/Dataset_Validation_2.csv')
```
